+++ Saab95.java
@@ -10,7 +10,7 @@ public class Saab95 extends Car {
 
 
 
-    public Saab95(int nrDoors, double enginePower, Color color, String modelName, double currentSpeed, double x, double y) {
+    public Saab95(int nrDoors, double enginePower, Color color, String modelName, double currentSpeed, int x, int y) {
         super(nrDoors, enginePower, color, modelName, currentSpeed, x, y);
         turboOn = false;
     }
