+++ Scania.java
@@ -5,7 +5,7 @@ public class Scania extends Truck {
     private final double platformRaiseAngle = 5;
     private double platformAngle;
 
-    public Scania(int nrDoors, double enginePower, Color color, String modelName, double currentSpeed, double x, double y) {
+    public Scania(int nrDoors, double enginePower, Color color, String modelName, double currentSpeed, int x, int y) {
         super(nrDoors, enginePower, color, modelName, currentSpeed, x, y);
         this.platformAngle = 0;
